move subscribe/2 and unsubscribe/2 to game_server
